# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Flexy GSM API
  description: API documentation for the Flexy GSM Backend
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
paths:
...
  /api/v1/users/{id}:
    get:
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: authorization
          in: header
          schema:
            type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    put:
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: authorization
          in: header
          schema:
            type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  example: any
                email:
                  example: any
                phone:
                  example: any
                wilaya:
                  example: any
                role:
                  example: any
                status:
                  example: any
                wallet:
                  example: any
                debt_limit:
                  example: any
                profit_percentage:
                  example: any
                permissions:
                  example: any
                password:
                  example: any
    delete:
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: authorization
          in: header
          schema:
            type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []